@@ -28,7 +28,7 @@
         }
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "ext-bcmath": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
@@ -36,7 +36,7 @@
         "eloquent/enumeration": "^6.0",
         "pragmarx/ia-str": "^7.0",
         "ramsey/uuid": "^4.0",
-        "somnambulist/collection": ">=4.0.1"
+        "somnambulist/collection": ">=5.0"
     },
     "require-dev": {
         "ext-pdo": "*",
